+++ GameServer/Room.h
@@ -11,9 +11,9 @@ public:
 public:
 	//	Recv Packet Process Func
 	void Enter(GameObjectRef gameObject);
-	void Leave(uint64 gameObjectId);
+	void Leave(uint32 gameObjectId);
 	void Move(Protocol::C_MOVE movePacket);
-	void Broadcast(SendBufferRef sendBuffer, uint64 ignoreId = 0);
+	void Broadcast(SendBufferRef sendBuffer, uint32 ignoreId = 0);
 
 public:
 	//	TEMP : Main Thread
@@ -24,7 +24,7 @@ public:
 	RoomRef GetRoomRef() { return static_pointer_cast<Room>(shared_from_this()); }
 
 private:
-	unordered_map<uint64, GameObjectRef> _gameObjects;
+	unordered_map<uint32, GameObjectRef> _gameObjects;
 };
 
 //	TEMP
